Split up src/sentry/models/integration.py
