+++ src/sentry/models/integrations/__init__.py
@@ -1,0 +1,20 @@
+__all__ = (
+    "DocIntegration",
+    "Integration",
+    "IntegrationExternalProject",
+    "OrganizationIntegration",
+    "PagerDutyService",
+    "ProjectIntegration",
+    "RepositoryProjectPathConfig",
+)
+
+# REQUIRED for migrations to run.
+from sentry.types.integrations import ExternalProviders  # NOQA
+
+from .doc_integration import DocIntegration
+from .integration import Integration
+from .integration_external_project import IntegrationExternalProject
+from .organization_integration import OrganizationIntegration
+from .pagerduty_service import PagerDutyService
+from .project_integration import ProjectIntegration
+from .repository_project_path_config import RepositoryProjectPathConfig
